test: manifest dynamically updates
