+++ apps/course/src/app.html
@@ -5,8 +5,6 @@
     <link rel="apple-touch-icon" sizes="180x180" href="%sveltekit.assets%/apple-touch-icon.png">
     <link rel="icon" type="image/png" sizes="32x32" href="%sveltekit.assets%/favicon-32x32.png">
     <link rel="icon" type="image/png" sizes="16x16" href="%sveltekit.assets%/favicon-16x16.png">
-    <link rel="manifest" href="/manifest.webmanifest">
-    <script async src="https://unpkg.com/pwacompat" crossorigin="anonymous"></script>
     <link rel="mask-icon" href="%sveltekit.assets%/safari-pinned-tab.svg" color="#37919b">
     <meta name="msapplication-TileColor" content="#ffffff">
     <meta name="theme-color" content="#ffffff">
